required 11 changes done
Change menu to Hide and Complain in trip history.
Show confirmation message on Hide in the menu in trip history.
If the string "finish" is coming, then it should be in green color; otherwise, it should be red.
Change font size to 20 and make the text bold in "Preferred Agencies".
Skip screen: Change Get Started button placement.
Add bullet points to the Welcome to Go Cab screen.
In the EULA screen, ensure the Continue button is in a scrollable view.
Change the message displayed on SMS tap in "Preferred Way to Receive SMS".
Change the message below the Verify button to match Android.
Show a message when entering the Main Screen, similar to Android.
Time Implementation: Currently in progress.